#287 - move FluentPanel to correct package
