+++ src/main/java/org/vaadin/viritin/layouts/MPanel.java
@@ -1,6 +1,6 @@
 package org.vaadin.viritin.layouts;
 
-import org.vaadin.viritin.ui.FluentPanel;
+import org.vaadin.viritin.fluency.ui.FluentPanel;
 
 import com.vaadin.ui.Component;
 import com.vaadin.ui.Panel;
