+++ src/main/java/org/vaadin/viritin/v7/fluency/ui/IPanel.java
@@ -1,6 +1,6 @@
 package org.vaadin.viritin.v7.fluency.ui;
 
-import org.vaadin.viritin.ui.FluentPanel;
+import org.vaadin.viritin.fluency.ui.FluentPanel;
 
 import com.vaadin.ui.Panel;
